feat(kad): add `mode` getter on `Behaviour` (#5573)
## Description

Small PR adding a getter for the `mode` attribute of the
`kad::Behaviour` in order to get the mode that the DHT is operating in,
at the moment.

In our project, we needed to expose an API endpoint which included the
mode that the DHT was operating. Having a getter was necessary so we are
upstreaming this change.

## Notes & open questions

<!--
Any notes, remarks or open questions you have to make about the PR which
don't need to go into the final commit message.
-->

## Change checklist

<!-- Please add a Changelog entry in the appropriate crates and bump the
crate versions if needed. See
<https://github.com/libp2p/rust-libp2p/blob/master/docs/release.md#development-between-releases>-->

- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] A changelog entry has been made in the appropriate crates