fix(gossipsub): Attempt to publish to at least mesh_n peers (#5578)
## Description

With flood published disabled we've noticed that it can be the case that
we have connected peers on topics but these peers are not in our mesh
(perhaps due to their own mesh requirements). Currently, we fail to
publish the message if there are no peers in our mesh.

This PR adjusts this logic to always attempt to publish to at least
mesh_n peers. If we have peers that are subscribed to a topic, we will
now attempt to publish messages to them (provided they have the required
score).

This PR also simplies the peer and respective topics by moving the topic
list each peer has subscribed to `PeerConnections` and removing both
`peer_topics` and `topic_peers` from the main `Behaviour`.
Per commit review is suggested.

---------

Co-authored-by: Darius Clark <dariusc93@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>